+++ R/sai.R
@@ -20,7 +20,7 @@ w_col <- function(source_unit, target) {
 ##' @param source a \code{sf} object - source spatial data.
 ##' @param source_unit a single \code{geometry} from the source dataset.
 ##' @param source_dt a \code{data.frame} object representing the source dataset
-##'     but excludying the \code{geometry}, i.e. the spatial information,
+##'     but excluding the \code{geometry}, i.e. the spatial information,
 ##'     column.
 ##' @param target a \code{sf} object - target spatial data.
 ##' @param var_vec a \code{numeric} vector with variances observed at the source
@@ -29,9 +29,9 @@ w_col <- function(source_unit, target) {
 ##' @param method a \code{character} representing the method to approximate the
 ##'     variance of the AI estimates. Possible values are "CS"
 ##'     (Cauchy-Schwartz) or "MI" (Moran's I).
-##' @param rho_mi \code{numeric} calcuated Moran's I.
+##' @param rho_mi \code{numeric} calculated Moran's I.
 ##' @return A \eqn{n \times m} \code{numeric} matrix. Where \eqn{n} is the
-##'     number of objservations in the target and \eqn{m} is the sample size in
+##'     number of observations in the target and \eqn{m} is the sample size in
 ##'     the source dataset.
 ##' @keywords internal
 build_w <- function(source, target) {
@@ -46,16 +46,18 @@ build_w <- function(source, target) {
   w_aux <- lapply(source, w_col,
                   target = target)
   w_aux <- do.call("cbind", w_aux)
-  div_target <- (1 / as.numeric(sf::st_area(target))) |>
-      diag()
+  div_target <- diag(
+      1 / as.numeric(sf::st_area(target))
+  )
 
   return(div_target %*% w_aux)
 }
 
 ##' @name weight_mat
 est_w <- function(W, source_dt, target) {
-    estimates <- (W %*% as.matrix(source_dt)) |>
-        as.data.frame()
+    estimates <- as.data.frame(
+        W %*% as.matrix(source_dt)
+    )
     cbind(target, estimates)
 }
 
@@ -70,10 +72,10 @@ morans_i <- function(sf_dt, variable) {
     stopifnot(inherits(variable, "character"))
     stopifnot(length(variable) == 1)
 
-    adj_mat <- sf::st_intersects(x = sf::st_geometry(sf_dt),
-                        sparse = FALSE) |>
-        as.numeric() |>
-        matrix(ncol = nrow(sf_dt))
+    adj_mat <- matrix(as.numeric(
+        sf::st_intersects(x = sf::st_geometry(sf_dt),
+                          sparse = FALSE)
+    ), ncol = nrow(sf_dt))
     
     diag(adj_mat) <- 0
     
@@ -114,9 +116,7 @@ var_w <- function(W, var_vec, target,
         cov_mat <- cov_mat * rho_mi
     }
 
-    se_est <- (W %*% tcrossprod(cov_mat, W)) |>
-        diag() |>
-        sqrt()
+    se_est <- sqrt(diag(W %*% tcrossprod(cov_mat, W)))
 
     return(transform(target, se_est = se_est))
 }
@@ -145,29 +145,6 @@ var_w <- function(W, var_vec, target,
 ##' 
 ##' @return the target (of type \code{sf}) with estimates of the variables
 ##'     observed at the source data.
-##'
-##' @examples
-##' data(nyc_surv)
-##' data(nyc_comd)
-##'
-##' ## creating variables that store the variance for each area
-##' ## this is done to exemplify the functionality of the package
-##' nyc_surv <- transform(nyc_surv,
-##'                       my_var = moe / qnorm(p = .975))
-##' nyc_surv <- transform(nyc_surv, my_var = my_var * my_var)
-##' 
-##' \dontrun{
-##' ## areal interpolation
-##' estimate_comd <-
-##'    ai(source = nyc_surv, target = nyc_comd,
-##'       vars = "estimate")
-##'
-##' ## areal interpolation with uncertainty estimation
-##' estimate_comd <-
-##'    ai_var(source = nyc_surv, target = nyc_comd,
-##'           vars = "estimate", vars_var = "my_var",
-##'           var_method = "MI")
-##' }
 ##' 
 ##' @export
 ai <- function(source, target,
@@ -224,10 +201,10 @@ ai_var <- function(source, target,
         } else {
             rho <- morans_i(source, vars)
         }
-        adj_mat <- sf::st_intersects(x = sf::st_geometry(source),
-                                     sparse = FALSE) |>
-            as.numeric() |>
-            matrix(ncol = nrow(source))
+        adj_mat <- matrix(as.numeric(
+            sf::st_intersects(x = sf::st_geometry(source),
+                              sparse = FALSE)
+        ), ncol = nrow(source))
         rho <- rho * adj_mat
         diag(rho) <- 1
         target <- var_w(W, source_dt[[vars_var]],
@@ -248,10 +225,10 @@ ai_var <- function(source, target,
         } else {
             rho <- morans_i(source, vars)
         }
-        adj_mat <- sf::st_intersects(x = sf::st_geometry(source),
-                                     sparse = FALSE) |>
-            as.numeric() |>
-            matrix(ncol = nrow(source))
+        adj_mat <- matrix(as.numeric(
+            sf::st_intersects(x = sf::st_geometry(source),
+                              sparse = FALSE)
+        ), ncol = nrow(source))
         rho2 <- rho * adj_mat
         diag(rho2) <- 1
         cov_mat1 <- rho * (tcrossprod(rep(1, .m)) - diag(.m))
@@ -263,13 +240,14 @@ ai_var <- function(source, target,
                       se_lb  = sqrt(diag(W %*% tcrossprod(cov_mat0, W))),
                       se_cs  = sqrt(diag(W %*% tcrossprod(cov_mat, W))),
                       se_mi  = sqrt(diag(W %*% tcrossprod(cov_mat1, W))),
-                      se_mi2 = sqrt(diag(W %*% tcrossprod(cov_mat2, W)))) |>
-            transform(se_itp = (rho * with(target, se_cs)) +
+                      se_mi2 = sqrt(diag(W %*% tcrossprod(cov_mat2, W))))
+        target <-
+            transform(target,
+                      se_itp = (rho * with(target, se_cs)) +
                           ((1 - rho) * with(target, se_lb)))
     } else {
         target <- var_w(W, source_dt[[vars_var]],
                         target, method = var_method)
     }
-    
     return(target)
 }