+++ NEWS.md
@@ -1,13 +1,13 @@
-# smile 1.0.1
+'# smile 1.0.1
 
 * Added a `NEWS.md` file to track changes to the package.
 * Fixed Spherical covariance function.
-* Added Cubic splince covariance function.
+* Added Cubic spline covariance function.
 
 # smile 1.0.2
 
 * added Wendland (1) covariance function
-* added "tappered" Matern covariance function
+* added "tapered" Matern covariance function
 
 # smile 1.0.3
 
@@ -29,3 +29,16 @@
 * Added icon
 * Improved documentation
 * Added examples
+
+# smile 1.0.4.2
+
+* New dataset included
+* Parallel version of `fit_spm2`
+
+# smile 1.0.5
+
+* Minor bug fixes and documentation enhancement.
+
+* Removed NYC datasets (they are too large).
+
+* Spatial areal interpolation vignette rewritten.